fix(ci): keep server_http_routes under line cap, refresh MCP tool golden
- move dev_submit_run + its env-cap/router helpers into server/server_dev_submit.c
  so server_http_routes.c stays under the 2500-line hard cap (link it into the
  two tests that pull server_http_routes.o)
- refresh the mcp_build_tools_list golden: +ensemble and +workflow_run tools,
  slimmed session family (count 50 -> 52)
- add 'ensembles' to the schema-sync DB1 table allowlist (was workflow_sessions)
- reformat with the pinned clang-format-19
